+++ data/panel.js
@@ -476,47 +476,25 @@ function newDeleteStreamButton(id, website){
 	
 	return node;
 }
-let I_am_watching_the_stream_of = "";
 function newShareStreamButton_onClick(event){
 	event.stopPropagation();
 	
 	let node = this;
+	let website = node.getAttribute("data-website");
 	let id = node.getAttribute("data-id");
 	let contentId = node.getAttribute("data-contentId");
-	let streamName = node.getAttribute("data-streamName");
-	let website = node.getAttribute("data-website");
-	let streamUrl = node.getAttribute("data-streamUrl");
-	let streamStatus = node.getAttribute("data-streamStatus");
 	
-	let facebookID = node.getAttribute("data-facebookID");
-	let twitterID = node.getAttribute("data-twitterID");
-	
-	let streamerAlias = streamName;
-	/*
-	if(facebookID != null && facebookID != ""){
-		
-	}*/
-	if(twitterID != null && twitterID != ""){
-		streamerAlias = `@${twitterID}`;
-		console.info(`${id}/${contentId} (${website}) twitter ID: ${twitterID}`)
-	}
-	
-	let shareMessage = `${I_am_watching_the_stream_of} ${streamerAlias}, "${streamStatus}"`;
-	console.info(shareMessage);
-	
-	// window.open(`https://twitter.com/intent/tweet?text=${encodeURIComponent(shareMessage)}&url=${streamUrl}&hashtags=LiveNotifier${(twitterID != "")? `&related=${twitterID}` : ""}&via=LiveNotifier`, '_blank');
-	window.open(`https://twitter.com/intent/tweet?text=${encodeURIComponent(shareMessage)}&url=${streamUrl}${(twitterID != "")? `&related=${twitterID}` : ""}&via=LiveNotifier`, '_blank');
+	my_port.sendData("shareStream", {
+		website: node.getAttribute("data-website"),
+		id: node.getAttribute("data-id"),
+		contentId: node.getAttribute("data-contentId"),
+	});
 }
 function newShareStreamButton(id, contentId, website, streamName, streamUrl, streamStatus, facebookID, twitterID){
 	let node = document.createElement("span");
+	node.setAttribute("data-website", website);
 	node.setAttribute("data-id", id);
 	node.setAttribute("data-contentId", contentId);
-	node.setAttribute("data-streamName", streamName);
-	node.setAttribute("data-website", website);
-	node.setAttribute("data-streamUrl", streamUrl);
-	node.setAttribute("data-streamStatus", streamStatus)
-	node.setAttribute("data-facebookID", facebookID);
-	node.setAttribute("data-twitterID", twitterID);
 	
 	let node_img =  document.createElement("i");
 	node_img.className = "material-icons";
@@ -837,9 +815,6 @@ chrome.runtime.onConnect.addListener(function(_port) {
 		switch(id){
 			case "initList":
 				initList(data);
-				break;
-			case "I_am_watching_the_stream_of":
-				I_am_watching_the_stream_of = data;
 				break;
 			case "updateOnlineCount":
 				listenerOnlineCount(data);